refactoring and split matchers into files
