+++ kotlintest-assertions/src/main/kotlin/io/kotlintest/matchers/reflection/propertyMatchers.kt
@@ -1,0 +1,6 @@
+package io.kotlintest.matchers.reflection
+
+import kotlin.reflect.KProperty
+
+inline fun <reified T> KProperty<*>.shouldBeOfType() = this.returnType.shouldBeOfType<T>()
+inline fun <reified T> KProperty<*>.shouldNotBeOfType() = this.returnType.shouldNotBeOfType<T>()
