+++ kotlintest-assertions/src/main/kotlin/io/kotlintest/matchers/reflection/typeMatchers.kt
@@ -1,0 +1,19 @@
+package io.kotlintest.matchers.reflection
+
+import io.kotlintest.Matcher
+import io.kotlintest.Result
+import io.kotlintest.should
+import io.kotlintest.shouldNot
+import kotlin.reflect.KType
+import kotlin.reflect.full.isSubtypeOf
+import kotlin.reflect.full.starProjectedType
+
+inline fun <reified T> KType.shouldBeOfType() = this should beOfType<T>()
+inline fun <reified T> KType.shouldNotBeOfType() = this shouldNot beOfType<T>()
+inline fun <reified T> beOfType() = object : Matcher<KType> {
+  override fun test(value: KType) = Result(
+      value.isSubtypeOf(T::class.starProjectedType),
+      "Type $value should be ${T::class}",
+      "Type $value should not be ${T::class}"
+  )
+}
